created refresh tokens functionality, modified file location etc., cleaned up a little
